Fix initial install code
Errors in the 2.2.2 install script were fixed which were causing
segfaults in some development environments. The fallback SQL INSERT code
has been modified to work in batches and to use the updated columns from
the table. In addition, the format of the  postcodes.csv file has been
changed to be usable by the new code.