Fix eParcel and postcode setup
Applied changes to the eParcel code to not use ternary operators, which
gets around the parser errors and (I think) makes it more readable.

Modified the postcode import to properly use LOAD DATA with the data
provided in the CSV file. This has been tested on several different
systems and should now work in the majority of cases, using faster LOAD
DATA where applicable or falling back to 18 INSERTs if it has to.
Both methods should be quick to run on most systems.

Version bumped to 2.2.5.

Fixes #5
Fixes #20