Add composer.json file
This makes use of the Magento-Packager extension to more easily create a
Magento extension package for distribution on Magento Connect.

At present this file requires the use of the updated Fontis version of
Magento-Packager as it includes multiple "magedesign" definitions, which
is not accounted for in the original Flagbit version.

https://github.com/fontis/Magento-Packager