Update eParcel payload for strict MySQL
Some MySQL servers will throw an exception in the way Magento inserts "null" data. Some servers will throw the following exception:

```
An error occurred while saving this configuration: SQLSTATE[23000]: Integrity constraint violation: 1048 Column 'stock_id' cannot be null
```

By passing an integer, we are sticking with the schema (which is a non-null integer). The alternative is to bump the version and make the field nullable.